REFACTO: set up TX Event Queue instead of counter, see below why.
There are no guarantee on the order in which ACK are received by the
initiator. As a consequence, op2 might be responsible for the increase
of the counter instead of op1. Since the completion mecanism relies on
a FIFO queue, op1 completion callback would be called too early. Thus,
full events are required to attach the operation related to the
completion event.