REFACTO: bugfix. consume unexpected header to avoid bad msg ordering.
All unexpected headers must be consumed when they are created. The
question is thus: when to consume them?
The answer should be: either directly because this means that a request
has been inserted in the unexpected list by UCP, thus the corresponding
receive zcopy will not be created. Or not at all because UCP tried to
cancel the receive which means the unexpected header has already been
consumed by BXI.