REFACTO: bugfix. bad handling of renvez-vous cancellation.
Rendezvous cancellation was resetting the INUSE operation flags and
putting back to the pool the OP. This was wrong if the PTL_EVENT_ACK
of the first rendezvous call was not processed before. Instead,
cancellation should also be handled using the completion counter.
Indeed, whether rndv is expected or unexpected, there will always be two
event to handle and thus a completion counter of 2.

Cancellation now just overwrite the handle so that, in case of
unexpected, either the PTL_EVENT_ACK will complete the OP or the
cancellation.