BXI: feat. support for offloaded unexpected rendezvous.
BXI supports matching in the overflow list. Using such feature provided
by Portals4, we are able to avoid performing the sw rendezvous when all
condition were here to offload the rendezvous. In such cases, the
unexpected rendezvous callback do not have to cancel anything and can
avoid performing the sw rendezvous.

TODO: to be documented.