BXI: bugfix. deadlock with sw rndv and offload rndv in recv.
Bug Description: When the sender selects a software rendezvous (SW RNDV)
despite the message size falling within the hardware rendezvous (HW
RNDV) range, the receiver offloads the get operation. If this operation
is later cancelled, the cancel call incorrectly returns UCS_INPROGRESS.
As a result, the remaining software protocol steps are not executed.

Fix: A workaround has been implemented to notify the BXI layer that the
cancelled receive originated from a software send. This ensures that
UCS_INPROGRESS is not returned in such cases.